refactor(telemetry): rename stats filter to aether_stats + add unit tests
Rename the proxy stats dynamic module from the path/target/contract name
`telemetry` / `aether_stats_filter` to `aether_stats`, and add Rust unit
tests for the filter logic.

Folder move:
  proxy/filters/telemetry -> proxy/filters/http/aether_stats

Contract rename (the `_filter` suffix is now implicit from the
filters/http/ path): the bazel target, Cargo package/lib, and the Envoy
dynamic-module basename all become `aether_stats`, so the proxy now loads
libaether_stats.so. The agent's statsModuleName is updated to match.
statsFilterName ("aether.stats") and the in-module dispatch key ("stats")
are unchanged.

Tests (//proxy/filters/http/aether_stats:aether_stats_test, 18 cases):
- classify_flag: every cause flag, empty->"-", ordering
- dest_service_from_cluster: mesh-suffix/port stripping, foreign names,
  empty/bare-suffix -> "unknown"
- ConfigData serde defaults / full parse / malformed rejection
- Filter hooks against the SDK's MockEnvoyHttpFilter: cluster capture,
  local-reply details, and the recorded label vector (attribute code,
  :status fallback, code-defaults-to-0, record-once idempotency)

Note: this is a breaking change to the proxy<->agent contract; the rebuilt
proxy image and the new agent must roll out together.
