+++ charts/agent/templates/proxy.yaml
@@ -58,7 +58,7 @@ spec:
       {{- end }}
       containers:
         - name: proxy
-          # Custom aether-proxy = stock Envoy base + baked aether_stats_filter
+          # Custom aether-proxy = stock Envoy base + baked aether_stats
           # module (//proxy:image_push). The module's search path is baked into
           # the image, so no volume or env is needed.
           image: {{ .Values.proxy.image.ref | quote }}
